+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+
+### Fixed
+
+- call the elements `focus` method inside of `focusable` to set focus
+
 ## [0.7.2] - 2018-07-23
 
 ### Fixed